introduce debug log and proper parsing of command line arguments
