+++ src/CMakeLists.txt
@@ -2,7 +2,7 @@ list(APPEND CMAKE_REQUIRED_DEFINITIONS)
 
 include_directories(${CMAKE_CURRENT_SOURCE_DIR} ${MMFD_BINARY_DIR} ${CMAKE_CURRENT_BINARY_DIR})
 
-add_executable(mmfd mmfd.c main.c babeld.c neighbour.c vector.c)
+add_executable(mmfd util.c main.c babeld.c neighbour.c vector.c)
 
 target_link_libraries(mmfd ${LIBNL_LIBRARIES} ${LIBNL_GENL_LIBRARIES})
